Woops: fixed incorrect function calls

+++ fib_cpp.cpp
@@ -9,7 +9,7 @@ int main(int argc, char* argv[]) {
 		return 1;
 	}
 	for(int i = 1; i < argc; i++) {
-		cout << fibonacci((double)atoi(argv[i])) << endl;
+		cout << fibonacci(atoi(argv[i])) << endl;
 	}
 	return 0;
 }
